requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0        # HTML/XML 解析器（比 html.parser 更快）
PyYAML>=6.0.1
python-dotenv>=1.0.0
